graph: targetmanifest/otpl: back out changes
- These changes attempted to fix offset detection, but were
  problematic.
- The right thing to do is abolish -repo and -offset flags and
  replace with a single -source flag (SourceLocation) that combines
  both.
- The reason for this is that when specifying -repo but not -offset
  we always assume a specific empty offset, so these flags interact
  in a confusing way.
- A single flag, to be either used or omitted would simplify things.
