+++ graph/otpldeploy.go
@@ -11,32 +11,22 @@ import (
 	"github.com/pkg/errors"
 )
 
-func newDetectedOTPLConfig(ls LogSink, wd LocalWorkDirShell, otplFlags *config.OTPLFlags, sc *sous.SourceContext) detectedOTPLDeployManifest {
+func newDetectedOTPLConfig(ls LogSink, wd LocalWorkDirShell, otplFlags *config.OTPLFlags) detectedOTPLDeployManifest {
 	if otplFlags.IgnoreOTPLDeploy {
 		return detectedOTPLDeployManifest{sous.NewManifests()}
 	}
-
 	otplParser := otpl.NewManifestParser(ls)
 	otplDeploySpecs, err := otplParser.ParseManifests(wd.Sh)
 	if err != nil {
 		// This is OK, we are detecting these speculatively.
 		return detectedOTPLDeployManifest{sous.NewManifests()}
 	}
-
-	ms := sous.NewManifests()
-
-	for _, man := range otplDeploySpecs.Snapshot() {
-		man.Source.Dir = sc.OffsetDir
-		ms.Add(man)
-	}
-
-	return detectedOTPLDeployManifest{ms}
+	return detectedOTPLDeployManifest{otplDeploySpecs}
 }
 
 func newUserSelectedOTPLDeploySpecs(
 	detected detectedOTPLDeployManifest,
 	tmid TargetManifestID,
-	sc *sous.SourceContext,
 	flags *config.OTPLFlags,
 	sm *ClientStateManager,
 	ls LogSink,
@@ -54,37 +44,38 @@ func newUserSelectedOTPLDeploySpecs(
 		}
 		return nowt, nil
 	}
-
-	if tmid.Source.Dir != sc.OffsetDir {
-		// TODO SS: Maybe support specifying other offsets eventually, for now
-		// it's hard to know the user's intention (via flags).
-		return nowt, fmt.Errorf("the offset of the current directory is %q; but you specified %q", sc.OffsetDir, tmid.Source.Dir)
-	}
-
 	mid := sous.ManifestID(tmid)
 	// we don't care about these flags when a manifest already exists
 	if _, ok := state.Manifests.Get(mid); ok {
-		return nowt, fmt.Errorf("manifest %s already exists", mid)
+		return nowt, nil
 	}
 
-	selected, err := getSelectedManifest(detected.Manifests, flags, ls)
-	if err != nil {
-		return nowt, err
+	var detectedManifest *sous.Manifest
+
+	if flavoredManifest, ok := detected.Manifests.Single(func(m *sous.Manifest) bool {
+		return m.Flavor == flags.Flavor
+	}); ok {
+		detectedManifest = flavoredManifest
+	} else {
+		flavors := detected.Manifests.Flavors()
+		if flags.Flavor == "" {
+			defer messages.ReportLogFieldsMessageToConsole("use the -flavor flag to pick a flavor", logging.WarningLevel, ls)
+		}
+		return nowt, fmt.Errorf("flavor %q not detected; pick from: %q", flags.Flavor, flavors)
 	}
 
-	if !flags.UseOTPLDeploy && !flags.IgnoreOTPLDeploy && len(selected.Deployments) != 0 {
+	if !flags.UseOTPLDeploy && !flags.IgnoreOTPLDeploy && len(detectedManifest.Deployments) != 0 {
 		return nowt, errors.New("otpl-deploy detected in config/, please specify either -use-otpl-deploy, or -ignore-otpl-deploy to proceed")
 	}
 	if !flags.UseOTPLDeploy {
 		return nowt, nil
 	}
-	if len(selected.Deployments) == 0 {
+	if len(detectedManifest.Deployments) == 0 {
 		return nowt, errors.New("use of otpl configuration was specified, but no valid deployments were found in config/")
 	}
 	deploySpecs := sous.DeploySpecs{}
-	for clusterName, spec := range selected.Deployments {
+	for clusterName, spec := range detectedManifest.Deployments {
 		if _, ok := state.Defs.Clusters[clusterName]; !ok {
-			// TODO SS: Error if clusters not recognised.
 			messages.ReportLogFieldsMessageToConsole(fmt.Sprintf("otpl-deploy config for cluster %q ignored", clusterName), logging.WarningLevel, ls)
 			continue
 		}
@@ -95,18 +86,7 @@ func newUserSelectedOTPLDeploySpecs(
 	}
 	// Detach the user selected from the detected manifest, in case something
 	// else relies on the detected ones.
-	selectedManifest := selected.Clone()
+	selectedManifest := detectedManifest.Clone()
 	selectedManifest.Deployments = deploySpecs
 	return userSelectedOTPLDeployManifest{selectedManifest}, nil
-}
-
-func getSelectedManifest(detected sous.Manifests, flags *config.OTPLFlags, ls logging.LogSink) (*sous.Manifest, error) {
-	if flavoredManifest, ok := detected.Single(func(m *sous.Manifest) bool { return m.Flavor == flags.Flavor }); ok {
-		return flavoredManifest, nil
-	}
-	flavors := detected.Flavors()
-	if flags.Flavor == "" {
-		defer messages.ReportLogFieldsMessageToConsole("use the -flavor flag to pick a flavor", logging.WarningLevel, ls)
-	}
-	return nil, fmt.Errorf("flavor %q not detected; pick from: %q", flags.Flavor, flavors)
 }
